+++ internal/roundtrip/roundtrip_test.go
@@ -8,7 +8,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/lex00/wetwire-honeycomb-go/internal/discovery"
+	"github.com/lex00/wetwire-honeycomb-go/internal/discover"
 	"github.com/lex00/wetwire-honeycomb-go/internal/serialize"
 	"github.com/lex00/wetwire-honeycomb-go/query"
 )